# backend/internal/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

type BearerAuth struct {
	Token string
	Roles []string
}

// GetToken returns the value of Token.
func (s *BearerAuth) GetToken() string {
	return s.Token
}

// GetRoles returns the value of Roles.
func (s *BearerAuth) GetRoles() []string {
	return s.Roles
}

// SetToken sets the value of Token.
func (s *BearerAuth) SetToken(val string) {
	s.Token = val
}

// SetRoles sets the value of Roles.
func (s *BearerAuth) SetRoles(val []string) {
	s.Roles = val
}

// GetUserinfoUnauthorized is response for GetUserinfo operation.
type GetUserinfoUnauthorized struct{}

func (*GetUserinfoUnauthorized) getUserinfoRes() {}

// Ref: #/components/schemas/IngressInfo
type IngressInfo struct {
	Name    string   `json:"name"`
	URL     string   `json:"url"`
	IconURL string   `json:"icon_url"`
	Tags    []string `json:"tags"`
}

// GetName returns the value of Name.
func (s *IngressInfo) GetName() string {
	return s.Name
}

// GetURL returns the value of URL.
func (s *IngressInfo) GetURL() string {
	return s.URL
}

// GetIconURL returns the value of IconURL.
func (s *IngressInfo) GetIconURL() string {
	return s.IconURL
}

// GetTags returns the value of Tags.
func (s *IngressInfo) GetTags() []string {
	return s.Tags
}

// SetName sets the value of Name.
func (s *IngressInfo) SetName(val string) {
	s.Name = val
}

// SetURL sets the value of URL.
func (s *IngressInfo) SetURL(val string) {
	s.URL = val
}

// SetIconURL sets the value of IconURL.
func (s *IngressInfo) SetIconURL(val string) {
	s.IconURL = val
}

// SetTags sets the value of Tags.
func (s *IngressInfo) SetTags(val []string) {
	s.Tags = val
}

// Ref: #/components/schemas/OIDCConfig
type OIDCConfig struct {
	Enabled  bool      `json:"enabled"`
	Issuer   OptString `json:"issuer"`
	ClientID OptString `json:"client_id"`
	Audience OptString `json:"audience"`
}

// GetEnabled returns the value of Enabled.
func (s *OIDCConfig) GetEnabled() bool {
	return s.Enabled
}

// GetIssuer returns the value of Issuer.
func (s *OIDCConfig) GetIssuer() OptString {
	return s.Issuer
}

// GetClientID returns the value of ClientID.
func (s *OIDCConfig) GetClientID() OptString {
	return s.ClientID
}

// GetAudience returns the value of Audience.
func (s *OIDCConfig) GetAudience() OptString {
	return s.Audience
}

// SetEnabled sets the value of Enabled.
func (s *OIDCConfig) SetEnabled(val bool) {
	s.Enabled = val
}

// SetIssuer sets the value of Issuer.
func (s *OIDCConfig) SetIssuer(val OptString) {
	s.Issuer = val
}

// SetClientID sets the value of ClientID.
func (s *OIDCConfig) SetClientID(val OptString) {
	s.ClientID = val
}

// SetAudience sets the value of Audience.
func (s *OIDCConfig) SetAudience(val OptString) {
	s.Audience = val
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/UserInfo
type UserInfo struct {
	Sub     string    `json:"sub"`
	Name    OptString `json:"name"`
	Email   OptString `json:"email"`
	Picture OptString `json:"picture"`
}

// GetSub returns the value of Sub.
func (s *UserInfo) GetSub() string {
	return s.Sub
}

// GetName returns the value of Name.
func (s *UserInfo) GetName() OptString {
	return s.Name
}

// GetEmail returns the value of Email.
func (s *UserInfo) GetEmail() OptString {
	return s.Email
}

// GetPicture returns the value of Picture.
func (s *UserInfo) GetPicture() OptString {
	return s.Picture
}

// SetSub sets the value of Sub.
func (s *UserInfo) SetSub(val string) {
	s.Sub = val
}

// SetName sets the value of Name.
func (s *UserInfo) SetName(val OptString) {
	s.Name = val
}

// SetEmail sets the value of Email.
func (s *UserInfo) SetEmail(val OptString) {
	s.Email = val
}

// SetPicture sets the value of Picture.
func (s *UserInfo) SetPicture(val OptString) {
	s.Picture = val
}

func (*UserInfo) getUserinfoRes() {}